Refresh PR execution footer after runs finish
