+++ reactor/tool.ts
@@ -3,6 +3,7 @@ import path from "node:path";
 import { execFile } from "node:child_process";
 import { promisify } from "node:util";
 import { loadConfig } from "./config";
+import { buildExecutionFooter, renderBodyWithExecutionFooter } from "./execution-footer";
 
 const execFileAsync = promisify(execFile);
 
@@ -436,14 +437,8 @@ function printHelp(): void {
 
 async function preparePullRequestBody(bodyPath: string, runDir: string): Promise<string> {
   const originalBody = await fs.readFile(bodyPath, "utf8");
-  if (originalBody.includes("<!-- openreactor:execution-footer -->")) {
-    return bodyPath;
-  }
   const footer = runDir ? await buildExecutionFooter(runDir) : "";
-
-  const nextBody = footer
-    ? `${originalBody.trimEnd()}\n\n${footer}\n`
-    : originalBody;
+  const nextBody = renderBodyWithExecutionFooter(originalBody, footer);
 
   if (nextBody === originalBody) {
     return bodyPath;
@@ -452,85 +447,6 @@ async function preparePullRequestBody(bodyPath: string, runDir: string): Promise
   const outputPath = path.join(path.dirname(bodyPath), `${path.basename(bodyPath, path.extname(bodyPath))}.openreactor${path.extname(bodyPath) || ".md"}`);
   await fs.writeFile(outputPath, nextBody, "utf8");
   return outputPath;
-}
-
-async function buildExecutionFooter(runDir: string): Promise<string> {
-  const runPath = path.join(runDir, "run.json");
-  type RunRecordSummary = {
-    triageExecution?: ExecutionSummary | null;
-    lastAgentExecution?: ExecutionSummary | null;
-    lastResult?: {
-      considerations?: string[] | null;
-    } | null;
-  };
-  let runRecord: RunRecordSummary | null = null;
-
-  try {
-    runRecord = JSON.parse(await fs.readFile(runPath, "utf8")) as RunRecordSummary;
-  } catch {
-    return "";
-  }
-
-  const lines = ["<!-- openreactor:execution-footer -->", "## OpenReactor Execution"];
-
-  if (runRecord?.triageExecution) {
-    lines.push(`- Triage: ${formatExecutionSummary(runRecord.triageExecution)}`);
-  }
-
-  if (runRecord?.lastAgentExecution) {
-    lines.push(`- Implementation: ${formatExecutionSummary(runRecord.lastAgentExecution)}`);
-  }
-
-  const considerations = (runRecord?.lastResult?.considerations ?? [])
-    .map((item: string) => item.trim())
-    .filter(Boolean)
-    .slice(0, 6);
-
-  if (considerations.length) {
-    lines.push("- Considered:");
-    lines.push(...considerations.map((item: string) => `  - ${item}`));
-  }
-
-  return lines.length > 2 ? lines.join("\n") : "";
-}
-
-interface ExecutionSummary {
-  providerLabel?: string | null;
-  model?: string | null;
-  reasoningEffort?: string | null;
-  serviceTier?: string | null;
-  toolLabel?: string | null;
-  durationMs?: number | null;
-}
-
-function formatExecutionSummary(execution: ExecutionSummary): string {
-  const parts = [
-    [execution.providerLabel, execution.model].filter(Boolean).join(" ").trim(),
-    execution.reasoningEffort ? `reasoning ${execution.reasoningEffort}` : "",
-    execution.serviceTier ? `service tier ${execution.serviceTier}` : "",
-    execution.toolLabel ? `tool ${execution.toolLabel}` : "",
-    typeof execution.durationMs === "number" ? `duration ${formatDurationMs(execution.durationMs)}` : ""
-  ].filter(Boolean);
-
-  return parts.join(" • ");
-}
-
-function formatDurationMs(durationMs: number): string {
-  const totalSeconds = Math.max(0, Math.round(durationMs / 1000));
-  const minutes = Math.floor(totalSeconds / 60);
-  const seconds = totalSeconds % 60;
-
-  if (minutes >= 60) {
-    const hours = Math.floor(minutes / 60);
-    const remainingMinutes = minutes % 60;
-    return `${hours}h ${remainingMinutes}m`;
-  }
-
-  if (minutes > 0) {
-    return `${minutes}m ${seconds}s`;
-  }
-
-  return `${totalSeconds}s`;
 }
 
 void main();
